requirements.txt
boto3==1.9.115
botocore==1.12.115
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.2
idna==2.5
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
MarkupSafe==1.1.1
newsapi-python==0.2.3
python-dateutil==2.8.0
requests==2.17.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.21.1
Werkzeug==0.14.1
